requirements.txt
discord.py
pystyle
discord-py-slash-command
asyncio
json
random
datetime